Add files via upload
This is Plot.2 and it follows after running a lag-time correction code that I've been working on. The red points seem to align along the zero lag value which is good, as we can see from Plot.1 where the points are drifting away from the 0 (zero) s value as days are passing by. 
The plot shows that I'm on the right way but several points are still not aligned along 0 s. This is an issue that I'm still trying to solve